% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2sql.R
\docType{package}
\name{csv2sql}
\alias{csv2sql}
\alias{csv2sql-package}
\title{A wrapper to convert csv files to sql files.}
\description{
This package converts csv files to sql files.
}
\author{
Jackson Kwok
}

