% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_sqlite.R
\name{csv_to_sqlite}
\alias{csv_to_sqlite}
\title{A function to convert csv file to sqlite3 file (using 'dplyr' package).}
\usage{
csv_to_sqlite(csv_name, db_name, table_name)
}
\arguments{
\item{csv_name}{Character string; path of the file.}

\item{db_name}{Character string; path of the database.}

\item{table_name}{Characeter string; (desired) table name.}
}
\description{
A function to convert csv file to sqlite3 file (using 'dplyr' package).
}

