% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_sqlite.R
\name{csv_to_sqlite_chunked}
\alias{csv_to_sqlite_chunked}
\title{A function to convert csv file to sqlite3 file with chunking (using 'dplyr' package).}
\usage{
csv_to_sqlite_chunked(csv_name, db_name, table_name, chunk_size = 10000)
}
\arguments{
\item{csv_name}{Character string; path of the file.}

\item{db_name}{Character string; path of the database.}

\item{table_name}{Characeter string; (desired) table name.}

\item{chunk_size}{Integer; the number of rows to include in each chunk.}
}
\description{
A function to convert csv file to sqlite3 file with chunking (using 'dplyr' package).
}

