% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_functions.R
\name{db_nrow}
\alias{db_nrow}
\title{Get number of rows of a database object}
\usage{
db_nrow(db_df0)
}
\arguments{
\item{A}{tbl object connected to a database.}
}
\description{
Get number of rows of a database object
}

